Split up styled components file
